rename internal to trigger
